+++ lib/honeybadger/integrations.rb
@@ -1,3 +1,4 @@
 module Honeybadger::Integrations; end
 
 require 'honeybadger/integrations/delayed_job'
+require 'honeybadger/integrations/sidekiq'